fix: CSB C040 CSBGRAPHICS_DAT exemption, MVE unsigned negation, DM2 cast, Theron NULL guard
CSB boot HUD handoff now accepts CSBGRAPHICS_DAT source and skips C040
validation when the resurrect surface is absent (CSBWin/Atari path).
MVE opcode 3 cast to int before negation to avoid unsigned wrap.
Stream parser opcode 0x08 minimum size harmonized to >= 6.
DM2 game load party coordinate cast corrected from int8_t to uint8_t.
Theron world hash NULL-checks its argument.

+++ src/dm2/dm2_v1_mve_video.c
@@ -146,7 +146,7 @@ static int mve_decode_block(DM2_V1_MveVideo *v, DM2_MveReader *r, uint8_t op, un
     case 0u:return mve_copy(v,v->last,bx,by,0,0);
     case 1u:return mve_copy(v,v->second_last,bx,by,0,0);
     case 2u:if(!mve_get8(r,&b))return 0;return b<56u?mve_copy(v,v->second_last,bx,by,8+(b%7u),b/7u):mve_copy(v,v->second_last,bx,by,-14+((b-56u)%29u),8+((b-56u)/29u));
-    case 3u:if(!mve_get8(r,&b))return 0;return b<56u?mve_copy(v,v->current,bx,by,-(8+(b%7u)),-(int)(b/7u)):mve_copy(v,v->current,bx,by,-(-14+(int)((b-56u)%29u)),-(8+(int)((b-56u)/29u)));
+    case 3u:if(!mve_get8(r,&b))return 0;return b<56u?mve_copy(v,v->current,bx,by,-(int)(8+(b%7u)),-(int)(b/7u)):mve_copy(v,v->current,bx,by,-(-14+(int)((b-56u)%29u)),-(8+(int)((b-56u)/29u)));
     case 4u:if(!mve_get8(r,&b))return 0;return mve_copy(v,v->last,bx,by,-8+(b&15u),-8+(b>>4u));
     case 5u:if(!mve_get8(r,&a)||!mve_get8(r,&b))return 0;return mve_copy(v,v->last,bx,by,(int)(int8_t)a,(int)(int8_t)b);
     case 6u:return 0;
